Fix/feedback_button_zakryva_jmeno_na_homescreen (#359) posunuti jmena a bodu dolu, aby se nekrylo s tlacitkem
